Move encoding into shared and xrepo.
