+++ lsif/src/shared/encoding/json.test.ts
@@ -1,15 +1,4 @@
-import { createFilter, testFilter, gzipJSON, gunzipJSON } from './encoding'
-
-describe('testFilter', () => {
-    it('should test set membership', async () => {
-        const filter = await createFilter(['foo', 'bar', 'baz'])
-        expect(await testFilter(filter, 'foo')).toBeTruthy()
-        expect(await testFilter(filter, 'bar')).toBeTruthy()
-        expect(await testFilter(filter, 'baz')).toBeTruthy()
-        expect(await testFilter(filter, 'bonk')).toBeFalsy()
-        expect(await testFilter(filter, 'quux')).toBeFalsy()
-    })
-})
+import { gzipJSON, gunzipJSON } from './json'
 
 describe('gzipJSON', () => {
     it('should preserve maps', async () => {
